requirements.txt
streamlit==1.50.0
pandas==2.3.3
plotly==6.3.1
supabase==2.22.0
requests==2.32.5
fpdf2==2.7.4
reportlab==4.0.7
qrcode==7.4.2
Pillow==10.1.0
python-dateutil==2.8.2
openpyxl==3.1.2
pyjwt==2.8.0